在通知栏上显示下载进度 使用Handler实现Service和Activity通信

+++ app/src/main/java/com/haocai/downloadservice/bean/ThreadInfo.java
@@ -10,9 +10,9 @@ package com.haocai.downloadservice.bean;
 public class ThreadInfo {
     private int id;
     private String url;
-    private int start;
+    private long start;
     private long end;
-    private int finished;
+    private long finished;
 
     public ThreadInfo(){
         super();
@@ -25,7 +25,7 @@ public class ThreadInfo {
      * @param end 每个线程下载的结束位
      * @param finish 每个线程已下载的结束位
      */
-    public ThreadInfo(int id, String url, int start, long end, int finish) {
+    public ThreadInfo(int id, String url, long start, long end, int finish) {
         this.id = id;
         this.url = url;
         this.start = start;
@@ -48,11 +48,11 @@ public class ThreadInfo {
         this.url = url;
     }
 
-    public int getStart() {
+    public long getStart() {
         return start;
     }
 
-    public void setStart(int start) {
+    public void setStart(long start) {
         this.start = start;
     }
 
@@ -64,11 +64,11 @@ public class ThreadInfo {
         this.end = end;
     }
 
-    public int getFinished() {
+    public long getFinished() {
         return finished;
     }
 
-    public void setFinished(int finished) {
+    public void setFinished(long finished) {
         this.finished = finished;
     }
